pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "yamlexcelconverter"
version = "1.0.0"
description = "Convert YAML ↔ Excel with GPG encryption support"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Paolo Cardamone"}
]
keywords = ["yaml", "excel", "converter", "gpg", "encryption"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml>=6.0",
    "openpyxl>=3.1.0",
    "tkinterdnd2>=0.3.0",
    "python-gnupg>=0.5.0",
]

[project.urls]
Homepage = "https://github.com/username/yamlconverter"
"Bug Reports" = "https://github.com/username/yamlconverter/issues"
"Source" = "https://github.com/username/yamlconverter"

[project.scripts]
yamlconverter = "yamlconverter.gui.main:main"

[project.gui-scripts]
yamlconverter-gui = "yamlconverter.gui.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
yamlconverter = ["../translations/*.json"]
